backend/requirements.txt
flask
pycryptodome
flask_cors
sqlite3